Update main.py
Added GARCH for volatility projections but was ultimately unable to make it compatible with numba so it's a little slow when using it. I fixed ARCH and GARCH to work with exp_rets in the several stocks class. I think I have to adjust ARCH slightly. Plan to add monte carlo sims to generate parameters for ARCH and GARCH as well. Also plan to use it to find the parameters for a heston model.